cbcvb
ccbcb

+++ dashboard.php
@@ -1,41 +1,59 @@
 <?php 
-	session_start();
+  session_start();
 
-      include_once('config.php');
+  include_once('config.php');
   
-      if (empty($_SESSION['username'])) {
-            header("Location: login.php");
-      }
+  if (empty($_SESSION['username'])) {
+    header("Location: login.php");
+  }
 
-      $sql = "SELECT * FROM users";
+  $sql = "SELECT * FROM users";
 
-      $selectUsers = $conn->prepare($sql);
-      $selectUsers->execute();
+  $selectUsers = $conn->prepare($sql);
+  $selectUsers->execute();
 
-      $users_data = $selectUsers->fetchAll();
+  $users_data = $selectUsers->fetchAll();
 ?>
 
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
-	<meta charset="utf-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="description" content="Dr PRO template project">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<link rel="stylesheet" type="text/css" href="styles/bootstrap-4.1.2/bootstrap.min.css">
-	<link href="plugins/font-awesome-4.7.0/css/font-awesome.min.css" rel="stylesheet" type="text/css">
-	<link rel="stylesheet" type="text/css" href="plugins/OwlCarousel2-2.2.1/owl.carousel.css">
-	<link rel="stylesheet" type="text/css" href="plugins/OwlCarousel2-2.2.1/owl.theme.default.css">
-	<link rel="stylesheet" type="text/css" href="plugins/OwlCarousel2-2.2.1/animate.css">
-	<link href="plugins/jquery-datepicker/jquery-ui.css" rel="stylesheet" type="text/css">
-	<link rel="stylesheet" type="text/css" href="styles/main_styles.css">
-	<link rel="stylesheet" type="text/css" href="styles/responsive.css">
-	<title>Index</title>
+  <meta charset="utf-8">
+  <meta content="width=device-width, initial-scale=1.0" name="viewport">
+  <title>EARTHTRADE</title>
+  <meta content="" name="description">
+  <meta content="" name="keywords">
+
+  <!-- Favicons -->
+  <link href="assets/img/favicon.png" rel="icon">
+  <link href="assets/img/apple-touch-icon.png" rel="apple-touch-icon">
+
+  <!-- Fonts -->
+  <link href="https://fonts.googleapis.com" rel="preconnect">
+  <link href="https://fonts.gstatic.com" rel="preconnect" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300;0,400;0,500;0,600;0,700;0,800;1,300;1,400;1,500;1,600;1,700;1,800&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Jost:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet">
+
+  <!-- Vendor CSS Files -->
+  <link href="assets/vendor/bootstrap/css/bootstrap.min.css" rel="stylesheet">
+  <link href="assets/vendor/bootstrap-icons/bootstrap-icons.css" rel="stylesheet">
+  <link href="assets/vendor/aos/aos.css" rel="stylesheet">
+  <link href="assets/vendor/glightbox/css/glightbox.min.css" rel="stylesheet">
+  <link href="assets/vendor/swiper/swiper-bundle.min.css" rel="stylesheet">
+
+  <!-- Main CSS File -->
+  <link href="assets/css/main.css" rel="stylesheet">
+
+  <!-- =======================================================
+  * Template Name: Arsha
+  * Template URL: https://bootstrapmade.com/arsha-free-bootstrap-html-template-corporate/
+  * Updated: May 10 2024 with Bootstrap v5.3.3
+  * Author: BootstrapMade.com
+  * License: https://bootstrapmade.com/license/
+  ======================================================== -->
 </head>
 <body>
-<?php 
-include_once('header.php');
-?>
+
 
 <main>
   <div class="d-flex justify-content-center align-items-center py-3 mb-3 border-bottom">
@@ -43,7 +61,24 @@ include_once('header.php');
   </div>
   <?php if($_SESSION['is_admin'] == 'true') { ?>
     <div class="container">
-      <h2>Users</h2>
+      <h2 class="text-center mb-3">Users</h2>
+      <div class="row mb-3 text-center">
+
+        <div class="col-3">
+        <a href="index.php" class='btn btn-darkgreen'>Go Home</a>
+        </div>    
+        <div class="col-3">
+         <a href="dashboard.php" class='btn btn-darkgreen'>Dashboard</a>
+        </div>    
+        <div class="col-3">
+          <a href="contactDashboard.php" class='btn btn-darkgreen'>Contact Inquiries</a>
+        </div>    
+        <div class="col-3">
+          <a href="plasticBR.php" class='btn btn-darkgreen'>Plastic Bottle Requests</a>
+        </div>   
+        
+
+      </div>
       <table class="table table-striped table-sm">
         <thead>
           <tr>
@@ -63,7 +98,7 @@ include_once('header.php');
             <td class="text-dark"><strong> <?php echo $user_data['username']; ?> </strong></td>
             <td class="text-dark"><strong> <?php echo $user_data['email']; ?>    </strong></td>
             <td> <a href="editUsers.php?id=<?= $user_data['id']; ?>">Update</a> </td>
-            <td> <a href="deleteUsers.php?id=<?= $user_data['id']; ?>">Delete</a> </td>
+            <td> <a href="deleteUsers.php?id=<?= $user_data['id']; ?>" onclick="return confirm('Are you sure you want to delete this user?')">Delete</a> </td>
           </tr>
           <?php } ?>
         </tbody>
@@ -106,5 +141,3 @@ include_once('header.php');
       <!-- <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script> -->
 </body>
 </html>
-
-
